requirements.txt
python-dotenv==1.0.0
googlemaps==4.10.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4 